Convert labels to a convention. Add h, t, x[bhw] commands.
GNU as doesn't warn one for many common things like jumping to a label
that doesn't exist due to a typo. So I redid the label naming convention
ti make it much easier to notice these problems. Now function entry points
begin with f_ and label in the bodied of functions are fb_ and always
associate upwards to the nearest f_. Macros use m_. I haven't done such
a thing with variables in memory or initialized data like strings yet.

The commands allow one to halt the processor, test reading a word into
a register from the repl, and to examine the contents of a single address or
range of address in a byte, half-word, or word sense.

Of course, more needs to be done, especially the implementation of backspace
and a little more error checking. It is possible for the processor to go off
into lala land duing error situations (which is ok since ultimately I'm
throwing all of this away and replacing it with a native lisp version).